Refactor Modal component and related files
