+++ packages/react/src/components/Modal/Modal.stories.tsx
@@ -32,10 +32,7 @@ export const Preview: StoryFn<typeof Modal> = (args) => {
         {...args}
         ref={modalRef}
       >
-        <Modal.Header
-          title='Modal header'
-          subtitle='Modal subtittel'
-        ></Modal.Header>
+        <Modal.Header subtitle='Modal subtittel'>Modal header</Modal.Header>
         <Modal.Content>
           <Paragraph>
             Lorem ipsum dolor sit, amet consectetur adipisicing elit. Blanditiis
@@ -62,19 +59,37 @@ export const CloseOnBackdropClick: StoryFn<typeof Modal> = () => {
         ref={modalRef}
         onInteractOutside={() => modalRef.current?.close()}
       >
-        <Modal.Header
-          title='Modal med closeOnBackdropClick og en veldig lang tittel'
-          subtitle='Her er det også divider'
-        ></Modal.Header>
-        <Divider color='subtle' />
+        <Modal.Header subtitle='Her er det også divider'>
+          Modal med closeOnBackdropClick og en veldig lang tittel
+        </Modal.Header>
         <Modal.Content>
           <Paragraph>
             Lorem ipsum dolor sit, amet consectetur adipisicing elit. Blanditiis
             doloremque obcaecati assumenda odio ducimus sunt et.
           </Paragraph>
         </Modal.Content>
-        <Divider color='subtle' />
         <Modal.Footer>Du kan også ha divider på footer</Modal.Footer>
+      </Modal>
+    </>
+  );
+};
+
+export const WithDivider: StoryFn<typeof Modal> = () => {
+  const modalRef = useRef<HTMLDialogElement>(null);
+
+  return (
+    <>
+      <Button onClick={() => modalRef.current?.showModal()}>Open Modal</Button>
+      <Modal ref={modalRef}>
+        <Modal.Header subtitle='Her er det også divider'>
+          Vi kan legge divider under header
+        </Modal.Header>
+        <Divider color='subtle' />
+        <Modal.Content>
+          <Paragraph>Rundt content</Paragraph>
+        </Modal.Content>
+        <Divider color='subtle' />
+        <Modal.Footer>Og over footer</Modal.Footer>
       </Modal>
     </>
   );
@@ -92,7 +107,7 @@ export const ModalWithForm: StoryFn<typeof Modal> = () => {
         onClose={() => setInput('')}
         width='450px'
       >
-        <Modal.Header title='Modal med skjema'></Modal.Header>
+        <Modal.Header>Modal med skjema</Modal.Header>
         <Modal.Content>
           <Textfield
             label='Navn'
