--- packages/react/src/components/Modal/ModalContent/index.ts
@@ -1,2 +1,0 @@
-import { ModalContent } from './ModaContent';
-export { ModalContent };
