--- packages/react/src/components/Modal/ModalContext.tsx
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-export type ModalContextProps = {
-  closeModal?: () => void;
-};
-
-export const ModalContext = createContext<ModalContextProps | null>(null);
