--- packages/react/src/components/Modal/ModalFooter/index.ts
@@ -1,2 +1,0 @@
-import { ModalFooter } from './ModalFooter';
-export { ModalFooter };
