--- packages/react/src/components/Modal/ModalHeader/index.ts
@@ -1,2 +1,0 @@
-import { ModalHeader } from './ModalHeader';
-export { ModalHeader };
